refactor(extract): house the Common Lisp extractor in extractors/
extract.py is being decomposed one language per module (#1212), so a
newly added language belongs in the package rather than adding another
500 lines to the file being emptied. The facade re-export and the
LANGUAGE_EXTRACTORS entry keep every existing importer and the registry
identity guard working unchanged.

The extractor body is unchanged; only the import plumbing moved.
